requirements.txt
Django==4.2
djangorestframework==3.14
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.14
pypdf==6.10.2
reportlab==4.4.10
